Created class server.ChatRoom
Implemented user to create a room by using the --room [room name] in the program
Created a new chatmessage type to parse command type messages.

Also updated the whoisin function to display the all the rooms and members
in the current room

Work in Progress:

-Functionality to chat within the room (broadcastToRooms() - server)
-GUI
-allowing certain users to create a room
-function so all users can join certain rooms

Usage:

when logged into the chat program use example "--room testroom" to create
a room called testroom. Once created the user that created it will be automatically
added to that room.

Signed-off-by: Jeff Wong <jwong@localhost.localdomain>